first test in the practice class

+++ app/src/main/java/package1/Bruch.java
@@ -87,7 +87,7 @@ public class Bruch implements Comparable<Bruch> {
      * @return Wert des Bruchs als Gleitkommazahl (double)
      */
     public double ausrechnen() {
-        return getZaehler() / getNenner();
+        return (double) getZaehler() / getNenner();
     }
 
     /**
